pyproject.toml
[project]
name = "git-viz"
version = "0.1.0"
description = "Browser-based git repository visualizer with interactive force-directed graphs"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn>=0.34.0",
	"gitpython>=3.1.0",
]

[dependency-groups]
dev = [
	"pytest>=8.0.0",
	"ruff>=0.9.0",
	"httpx>=0.28.0",
]

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
